% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bind_data.R
\name{bind_data}
\alias{bind_data}
\title{Merge Data Sets}
\usage{
bind_data(s1, s2, s0 = NULL)
}
\arguments{
\item{s1}{The first phase \code{\link{data.frame}}.}

\item{s2}{The second phase \code{\link{data.frame}}.}

\item{s0}{Possibly a null phase \code{\link{data.frame}}.}
}
\value{
A \code{\link{data.frame}} with added boolean variables \code{phase1}
and \code{phase2}.
}
\description{
You will usually have different data sets for the first, second and possibly
a null phase. This functions binds them into a data sets suitable to create
an object of class \code{\linkS4class{saeObj}}.
}
\note{
This is experimental!
}
\examples{
data(list= paste0("s", 1:2), package = "maSAE")
str(s1)
str(s2)
str(bind_data(s1, s2))
}
\keyword{internal}
