% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{s0}
\alias{s0}
\title{Example s0 Data Set}
\format{
A data frame with 9008 observations on the following 6 variables.
}
\usage{
data(s0, package = "maSAE")
}
\description{
Artificial null phase sampling data used for examples in \pkg{maSAE}.
}
\details{
\describe{
   \item{\code{clustid}}{See \code{"\link[=s2]{?maSAE::s2}"}}
   \item{\code{x1}}{See \code{"\link[=s2]{?maSAE::s2}"}}
   \item{\code{x2}}{See \code{"\link[=s2]{?maSAE::s2}"}}
   \item{\code{x3}}{See \code{"\link[=s2]{?maSAE::s2}"}}
   \item{\code{inclusion}}{See \code{"\link[=s2]{?maSAE::s2}"}}
   \item{\code{g}}{See \code{"\link[=s2]{?maSAE::s2}"}}
 }
}
\keyword{datasets}
