% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\docType{class}
\name{savObj-class}
\alias{savObj-class}
\title{Small Area Virtual Object Class}
\description{
Common slots for classes \code{sadObj} and \code{saeObj}.
}
\section{Slots}{

\describe{
\item{\code{data}}{See \code{"\linkS4class{saeObj}"}.}

\item{\code{f}}{See \code{"\linkS4class{saeObj}"}.}

\item{\code{cluster}}{See \code{"\linkS4class{saeObj}"}.}

\item{\code{include}}{See \code{"\linkS4class{saeObj}"}.}
}}

\note{
The slots are described in
\code{"\link[=saeObj-class]{class?maSAE::saeObj}"}, since this is the main class
of the package.
}
\section{Objects from the Class}{
 A virtual Class: No objects may be created
from it.
}

\examples{

showClass("savObj")
}
\seealso{
\code{"\link[stats:formula]{?stats::formula}"},
\code{"\link[=sadObj-class]{class?maSAE::sadObj}"} and
\code{"\link[=saeObj-class]{class?maSAE::saeObj}"}.

Other classes: 
\code{\link{characterOrNULL-class}},
\code{\link{sadObj-class}},
\code{\link{saeObj-class}}
}
\concept{classes}
\keyword{classes}
\keyword{internal}
